# post man import/api-docs (1).json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/user/addUser":{"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/transactions/createTransaction":{"post":{"tags":["transaction-controller"],"operationId":"createTransaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionResponseDto"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtAuthResponseDto"}}}}}}},"/user/updateActiveStatus/{userId}/status":{"patch":{"tags":["user-controller"],"operationId":"updateActiveStatus","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"active","in":"query","required":true,"schema":{"type":"string","enum":["ACTIVE","CLOSED","SUSPENDED","BLOCKED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/user/getUserByName/{userName}":{"get":{"tags":["user-controller"],"operationId":"getUserByName","parameters":[{"name":"userName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/user/getUserById/{userId}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/user/getAllUsers":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}}},"/transactions/getTransactionById/{transactionId}":{"get":{"tags":["transaction-controller"],"operationId":"getTransactionById","parameters":[{"name":"transactionId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionResponseDto"}}}}}}},"/transactions/getAllTransactionsByUserId":{"get":{"tags":["transaction-controller"],"operationId":"getAllTransactionsByUserId","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"createdAt,desc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageTransactionResponseDto"}}}}}}},"/user/deleteUserById/{userId}":{"delete":{"tags":["user-controller"],"operationId":"deleteUserById","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UserRequestDto":{"required":["email","passwordHash","phoneNumber","role","userName"],"type":"object","properties":{"userName":{"type":"string"},"passwordHash":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"pattern":"^[0-9]{10}$","type":"string"},"role":{"type":"string","enum":["ADMIN","USER"]}}},"UserResponseDto":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"userName":{"type":"string"},"isActive":{"type":"string","enum":["ACTIVE","CLOSED","SUSPENDED","BLOCKED"]}}},"TransactionRequestDto":{"required":["amount","channel","country","deviceId","sourceAccount","transactionType"],"type":"object","properties":{"amount":{"type":"number"},"transactionType":{"type":"string","enum":["WITHDRAWAL","DEPOSIT","TRANSFER","SELF"]},"sourceAccount":{"pattern":"^ACC-[A-Z]{2}-[0-9]{6,12}$","type":"string"},"targetAccount":{"pattern":"^ACC-[A-Z]{2}-[0-9]{6,12}$","type":"string"},"channel":{"type":"string","enum":["UPI","CARD","NET_BANKING"]},"country":{"pattern":"^[A-Z]{2}$","type":"string"},"deviceId":{"pattern":"^DEV-(ANDROID|IOS|WEB)-[A-Z0-9]{5,20}$","type":"string"}}},"TransactionResponseDto":{"type":"object","properties":{"transactionId":{"type":"string","format":"uuid"},"amount":{"type":"number"},"sourceAccount":{"type":"string"},"targetAccount":{"type":"string"},"transactionType":{"type":"string","enum":["WITHDRAWAL","DEPOSIT","TRANSFER","SELF"]},"channel":{"type":"string","enum":["UPI","CARD","NET_BANKING"]},"createdAt":{"type":"string","format":"date-time"},"transactionStatus":{"type":"string","enum":["INITIATED","DECLINED","APPROVED","REVIEW"]}}},"UserLoginRequest":{"required":["password","phoneNumber"],"type":"object","properties":{"phoneNumber":{"pattern":"^[0-9]{10}$","type":"string"},"password":{"type":"string"}}},"JwtAuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"tokenType":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"},"issuedAt":{"type":"integer","format":"int64"}}},"PageTransactionResponseDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponseDto"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}}}}}